ui: unify error modal + feedback lane; refactor error helpers; polish signup/connected apps
